+++ lib/minitest_to_rspec/processor.rb
@@ -6,22 +6,23 @@ require_relative "subprocessors/iter"
 
 module MinitestToRspec
   class Processor < SexpProcessor
-    def initialize(rails)
+    def initialize(rails, mocha)
       super()
       self.strict = false
+      @mocha = mocha
       @rails = rails
     end
 
     def process_call(exp)
-      Subprocessors::Call.new(exp, @rails).process
+      Subprocessors::Call.new(exp, @rails, @mocha).process
     end
 
     def process_class(exp)
-      Subprocessors::Klass.new(exp, @rails).process
+      Subprocessors::Klass.new(exp, @rails, @mocha).process
     end
 
     def process_iter(exp)
-      Subprocessors::Iter.new(exp).process
+      Subprocessors::Iter.new(exp, @rails, @mocha).process
     end
   end
 end